moved `shallowEqual` and `is` utilities
