+++ src/observer.js
@@ -5,7 +5,7 @@ import { findDOMNode as baseFindDOMNode } from "react-dom"
 
 import EventEmitter from "./utils/EventEmitter"
 import inject from "./inject"
-import { patch as newPatch, newSymbol } from "./utils/utils"
+import { patch as newPatch, newSymbol, shallowEqual } from "./utils/utils"
 
 const mobxAdminProperty = $mobx || "$mobx"
 const mobxIsUnmounted = newSymbol("isUnmounted")
@@ -98,32 +98,6 @@ export const errorsReporter = new EventEmitter()
 
 function patch(target, funcName) {
     newPatch(target, funcName, reactiveMixin[funcName])
-}
-
-function shallowEqual(objA, objB) {
-    //From: https://github.com/facebook/fbjs/blob/c69904a511b900266935168223063dd8772dfc40/packages/fbjs/src/core/shallowEqual.js
-    if (is(objA, objB)) return true
-    if (typeof objA !== "object" || objA === null || typeof objB !== "object" || objB === null) {
-        return false
-    }
-    const keysA = Object.keys(objA)
-    const keysB = Object.keys(objB)
-    if (keysA.length !== keysB.length) return false
-    for (let i = 0; i < keysA.length; i++) {
-        if (!hasOwnProperty.call(objB, keysA[i]) || !is(objA[keysA[i]], objB[keysA[i]])) {
-            return false
-        }
-    }
-    return true
-}
-
-function is(x, y) {
-    // From: https://github.com/facebook/fbjs/blob/c69904a511b900266935168223063dd8772dfc40/packages/fbjs/src/core/shallowEqual.js
-    if (x === y) {
-        return x !== 0 || 1 / x === 1 / y
-    } else {
-        return x !== x && y !== y
-    }
 }
 
 function makeComponentReactive(render) {
